fix back navigation on redirect
